gcodeserver/requirements.txt
PyDispatcher
PySerial
